+++ packages/sage-react/lib/Modal/ModalHeader.jsx
@@ -1,13 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
+import { Card } from '../Card';
+import { Icon } from '../Icon';
+import { Popover } from '../Popover';
 import { ModalHeaderAside } from './ModalHeaderAside';
-import { SageClassnames } from '../configs';
+import { SageClassnames, SageTokens } from '../configs';
 
 export const ModalHeader = ({
   aside,
   children,
   className,
+  popover,
+  icon,
+  image,
+  subheader,
   title,
   ...rest
 }) => {
@@ -18,9 +25,37 @@ export const ModalHeader = ({
 
   return (
     <header className={classNames} {...rest}>
-      {title && <h5 className={SageClassnames.TYPE.HEADING_4}>{title}</h5>}
-      {children}
-      {aside && <ModalHeaderAside>{aside}</ModalHeaderAside>}
+      <div className="sage-modal__header-text">
+        <Card.Row
+          gridTemplate={
+            icon || image
+              ? SageTokens.GRID_TEMPLATES.ETE
+              : SageTokens.GRID_TEMPLATES.TE
+          }
+        >
+          {icon && <Icon className="sage-modal__header-icon" icon={icon.name} color={icon.color} /> }
+          {image && <img src={image.src} alt={image.alt || ''} />}
+          {title && <h1 className={SageClassnames.TYPE.HEADING_4}>{title}</h1>}
+          {aside && <ModalHeaderAside>{aside}</ModalHeaderAside>}
+        </Card.Row>
+        {subheader && (
+          <Card.Row>
+            <h2 className={SageClassnames.TYPE.HEADING_6}>{subheader}</h2>
+            {popover && (
+              <Popover
+                title={popover.title}
+                iconOnly={true}
+                icon={SageTokens.ICONS.QUESTION_CIRCLE}
+                moreLinkText={popover.linkText}
+                moreLinkURL={popover.link}
+              >
+                {popover.content}
+              </Popover>
+            )}
+          </Card.Row>
+        )}
+        {children}
+      </div>
     </header>
   );
 };
@@ -31,6 +66,10 @@ ModalHeader.defaultProps = {
   aside: null,
   children: null,
   className: null,
+  icon: null,
+  image: null,
+  popover: null,
+  subheader: null,
   title: null,
 };
 
@@ -38,5 +77,20 @@ ModalHeader.propTypes = {
   aside: PropTypes.node,
   children: PropTypes.node,
   className: PropTypes.string,
+  icon: PropTypes.shape({
+    color: PropTypes.oneOf(Object.values(Icon.COLORS)),
+    name: PropTypes.oneOf(Object.values(Icon.ICONS)),
+  }),
+  image: PropTypes.shape({
+    alt: PropTypes.string,
+    src: PropTypes.string
+  }),
+  popover: PropTypes.shape({
+    content: PropTypes.node,
+    link: PropTypes.string,
+    linkText: PropTypes.string,
+    title: PropTypes.string
+  }),
+  subheader: PropTypes.string,
   title: PropTypes.string,
 };