+++ packages/sage-react/CHANGELOG.md
@@ -3,6 +3,42 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.56.0](https://github.com/Kajabi/sage-lib/compare/@kajabi/sage-react@0.55.0...@kajabi/sage-react@0.56.0) (2021-09-01)
+
+
+### Bug Fixes
+
+* **icons:** menu rename ([aa3fa40](https://github.com/Kajabi/sage-lib/commit/aa3fa40e1dbfad70a3d87d209c40f8a3fd97518e))
+* **icons:** menu rename ([47660cb](https://github.com/Kajabi/sage-lib/commit/47660cbd08f422725946abf5a34e358c3e21435f))
+* **icons:** menu rename ([74490f0](https://github.com/Kajabi/sage-lib/commit/74490f06829ec549f6c8dda4f21bd24b8f66c4fa))
+* **icons:** menu rename ([4b7c6b3](https://github.com/Kajabi/sage-lib/commit/4b7c6b33810a35f52b11f45a1ec8f25c6f4ea37f))
+* **icons:** updated icon names ([7aed9cc](https://github.com/Kajabi/sage-lib/commit/7aed9cc6b893d8e98459745ed2cb88c0a8accf14))
+* **icons:** updated icon names ([e496010](https://github.com/Kajabi/sage-lib/commit/e4960105e1d07f617dc004cee3ba7e71cdc14f56))
+* **icons:** updated icon names ([6f607f2](https://github.com/Kajabi/sage-lib/commit/6f607f21006e8eeede86270fad0f3e365434bddf))
+* **icons:** updated icon names ([625e3a1](https://github.com/Kajabi/sage-lib/commit/625e3a1df47a160c6bd597fa96af394a9bc85403))
+
+
+### Features
+
+* **icon:** add adjacentType on React component ([f7731f7](https://github.com/Kajabi/sage-lib/commit/f7731f70b4e569e262d4b551dca086c936cf6d54))
+* **icons:** added new icons ([6f8796d](https://github.com/Kajabi/sage-lib/commit/6f8796d4c54c127cbf74156158bb8e5a56d3418a))
+* **icons:** added new icons ([dcffc68](https://github.com/Kajabi/sage-lib/commit/dcffc687e94bcb3cf9a0af0a860da50c24a0bc56))
+* **icons:** added new icons ([d16fb8f](https://github.com/Kajabi/sage-lib/commit/d16fb8f23d17ce8cd81c92071db3d5334368c8e5))
+* **icons:** added new icons ([a2faba6](https://github.com/Kajabi/sage-lib/commit/a2faba6673c12041762c663295fdcd87ded56116))
+* **media-tile:** add react component ([cf845ba](https://github.com/Kajabi/sage-lib/commit/cf845babee6657f9ec9a316d6c06a345c82905f0))
+* **media-tile:** further refine react component and add group component and stories ([aab6f9d](https://github.com/Kajabi/sage-lib/commit/aab6f9dd831fa536a25b5e1c9b86101f97f56153))
+* **modal:** updates React component and documentation ([e508d93](https://github.com/Kajabi/sage-lib/commit/e508d93d80e30db73413259f02a00d208f684d87))
+* **panel:** added custom class for PanelSubheading components ([4340bc0](https://github.com/Kajabi/sage-lib/commit/4340bc094784dbd8307ceb08e90d0c15617e347e))
+* **panel:** added custom class logic to PanelHeader React component ([888a729](https://github.com/Kajabi/sage-lib/commit/888a729b7073e0eeb07cb4c38cbb21bf098cf21d))
+* **panel:** added custom class when children are present in Rails component ([a7fde7d](https://github.com/Kajabi/sage-lib/commit/a7fde7d37b59233711b4589896383ea967d5eb50))
+* **panel:** added subtext React component, story context ([c66f311](https://github.com/Kajabi/sage-lib/commit/c66f3114b1d566ab6877de836a52904e92c8cd61))
+* **panel:** added subtext to subheader React component ([4c04ff3](https://github.com/Kajabi/sage-lib/commit/4c04ff37b5372f7b57b50541ff33ba08ea6b8c1f))
+* **panel-figure:** add latest media configurations to react component ([4ea615d](https://github.com/Kajabi/sage-lib/commit/4ea615d08a860eb05c336fe9efb2564ab5e2bdab))
+
+
+
+
+
 # [0.55.0](https://github.com/Kajabi/sage-lib/compare/@kajabi/sage-react@0.54.7...@kajabi/sage-react@0.55.0) (2021-08-25)